Report_qa_app/requirements.txt
streamlit>=1.37.0
python-docx>=0.8.11
pdfplumber>=0.11.0
pyyaml>=6.0.1
pandas>=2.2.2
regex>=2024.5.15
pytest>=7.0
regex>=2024.5.15
pytest>=7.0
numpy>=1.24
openai>=1.0.0